refactor network name and password input into two separate steps
